quiver/requirements.txt
APScheduler==3.0.2
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
configparser==5.0.2
convcolors==2.2.0
cryptography==3.4.7
Django==3.2.5
django-base64field==1.0
django-cors-headers==3.7.0
django-model-utils==4.1.1
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-spectacular==0.17.2
extcolors==1.0.0
idna==2.10
inflection==0.5.1
jsonschema==3.2.0
numpy==1.23.2
pandas==1.4.3
pdfkit==1.0.0
Pillow==8.3.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sentry-sdk==1.5.8
six==1.16.0
spintax==1.0.4
sqlparse==0.4.1
traitlets==5.4.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6 